GameServer: компиляция под POSIX пройдена полностью
Все 65 исходников GameServer компилируются clang на macOS ARM64. Сборка
доходит до линковки.

Расширен PlatformCompat.h: соглашения о вызовах (WINAPI), консоль через
ANSI-escape вместо SetConsoleTextAttribute, заглушки цикла сообщений,
SEH-типы, MessageBox в stderr, OutputDebugString, MSVC-функции строк
(strncpy_s, strncat_s, _snprintf_s, strlwr, itoa, _countof).

OdbcCompat.h решает конфликт типов с unixODBC: его ULONG это unsigned long,
то есть 64 бита на POSIX, а наш строго 32-битный. Через
ALREADY_HAVE_WINDOWS_TYPE подавляем определения unixODBC и доопределяем
недостающее — сделать наш ULONG 64-битным нельзя, тип встречается в
структурах с фиксированной раскладкой.

Смена компилятора вскрыла реальные дефекты, а не только различия платформ:

1. Передача std::string в printf-подобные функции с переменным числом
   аргументов (4 места). Это UB: вместо указателя на символы уезжают байты
   объекта. На MSVC 'работало' случайно из-за SSO. Добавлен .c_str().

2. friend-объявление не делает имя видимым для обычного поиска — только для
   ADL. Entity.h использовал SubMap* после friend class SubMap;.

3. C_PRINT разворачивался в printf(s, ) при вызове без переменных
   аргументов. Исправлено через __VA_OPT__.

4. #include <list> ниже места использования; register в C++17.

SystemDialog.cpp исключён из POSIX-сборки: это чистый Win32-диалог.
<stacktrace> подключается через __has_include — в libc++ на macOS его нет,
в libstdc++ на Linux есть.

+++ sources/Server/GameServer/src/Character/Character.cpp
@@ -354,7 +354,9 @@ void CCharacter::SwitchMap(SubMap *pCSrcMap, const char *szTarMapName, std::int3
 			m_SFightInit.chTarType = 0;
 			m_CChaAttr.Init(GetCat());
 			Corsairs::Util::Square	SSrcShape = GetShape();
-			Corsairs::Util::Square	STarShape = {{lTarX, lTarY}, GetRadius()};
+			// Явное приведение: на POSIX long шире int32, и сужение в списке
+			// инициализации — ошибка, а не предупреждение.
+			Corsairs::Util::Square	STarShape = {{static_cast<std::int32_t>(lTarX), static_cast<std::int32_t>(lTarY)}, static_cast<std::int32_t>(GetRadius())};
 			if (!pCSrcMap->Enter(&STarShape, this))
 				pCSrcMap->Enter(&SSrcShape, this);
 		}
